+++ post-fs-data.sh
@@ -1,0 +1,3 @@
+#!/system/bin/sh
+# early boot hook: clear Android OTA font directory
+rm -rf /data/fonts